This adds to new github actions.
The main action `release.yaml` cuts a release when new `vXYZ` tags are pushed and attaches the resolved `ko` yaml to the release as `release.yaml`.

The other action `latest.yaml` runs the same workflow whenever `master` is updated, but stops short of creating a release.  How we surface the resolved yaml this produces is TBD.