Add configurable checkpoints
This change allows to configure the checkpointing behavior in the spec
of the vSphere source. If no or the zero value for periodSeconds is
specified, a default (10s) will be applied. If maxAgeSeconds is 0,
replay is disabled. Optionally, this behavior can be configured via the
kn plugin. Thus, two new flags are introduced, checkpoint-age and
checkpoint-period.

Signed-off-by: Michael Gasch <mgasch@vmware.com>