Make data encoding configurable
This change adds support for configurable CloudEvent dataencoding
schemes. Currently, only application/xml and application/JSON are
supported.

The default encoding (if unspecified) remains application/xml to
preserve backwards-compatibility.

The kn plugin has a new flag to configure this behavior. For ease of
use, "xml" or "json" are accepted, i.e. "application/" must not be
specified.

Closes: #222
Signed-off-by: Michael Gasch <mgasch@vmware.com>